docs: rename docs dir to match component name
