+++ packages/components/src/ButtonGroup/_docs/ButtonGroup.stories.tsx
@@ -4,9 +4,7 @@ import { FilterButtonBase } from '~components/Filter/FilterButton/subcomponents/
 import { ButtonGroup } from '..'
 
 const meta = {
-  // Not to be nested until full KAIO migration
-  // title: "Components/Button/Button Group",
-  title: 'Components/Button Group',
+  title: 'Components/ButtonGroup',
   component: ButtonGroup,
   args: { children: undefined },
 } satisfies Meta<typeof ButtonGroup>
