+++ packages/components/src/Filter/FilterButton/_docs/FilterButtonRemovable.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { FilterButtonRemovable } from '../FilterButtonRemovable'
 
 const meta = {
-  title: 'Components/Filter Base/Filter Buttons/Filter Button Removable',
+  title: 'Components/Filter Base/Filter Buttons/FilterButtonRemovable',
   component: FilterButtonRemovable,
   args: {
     triggerButtonProps: {
