+++ packages/components/src/Filter/FilterDatePicker/_docs/FilterDatePicker.stickersheet.stories.tsx
@@ -11,7 +11,7 @@ import { FilterDatePickerField } from '../subcomponents/FilterDatePickerField'
 const IS_CHROMATIC = isChromatic()
 
 export default {
-  title: 'Components/Filter Date Picker',
+  title: 'Components/FilterDatePicker',
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
