+++ packages/components/src/Modal/GenericModal/_docs/GenericModal.spec.stories.tsx
@@ -6,7 +6,7 @@ import { expect, userEvent, within, waitFor } from '@storybook/test'
 import { GenericModal, ModalAccessibleLabel, ModalBody, ModalHeader } from '../index'
 
 const meta: Meta<typeof GenericModal> = {
-  title: 'Components/Modals/Generic Modal/Tests',
+  title: 'Components/Modals/GenericModal/Tests',
   component: GenericModal,
 }
