+++ packages/components/src/Modal/GenericModal/_docs/GenericModal.stories.tsx
@@ -3,7 +3,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { GenericModal } from '../index'
 
 const meta = {
-  title: 'Components/Modals/Generic Modal',
+  title: 'Components/Modals/GenericModal',
   component: GenericModal,
   args: {
     children: 'Example content',
