+++ packages/components/src/TextArea/_docs/TextArea.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { TextArea } from '../index'
 
 const meta = {
-  title: 'Components/Text Input controls/Text Area',
+  title: 'Components/Text Input controls/TextArea',
   component: TextArea,
   parameters: {
     a11y: {
