+++ packages/components/src/TextField/_docs/TextField.stories.tsx
@@ -4,7 +4,7 @@ import { Icon } from '~components/__future__/Icon'
 import { TextField } from '../index'
 
 const meta = {
-  title: 'Components/Text Input controls/Text Field',
+  title: 'Components/Text Input controls/TextField',
   component: TextField,
   args: {
     labelText: 'Label',
