+++ packages/components/src/VisuallyHidden/_docs/VisuallyHidden.stories.tsx
@@ -3,7 +3,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { VisuallyHidden } from '../index'
 
 const meta = {
-  title: 'Components/Visually Hidden',
+  title: 'Components/VisuallyHidden',
   component: VisuallyHidden,
   args: {
     children: '👋 Hello!',
