+++ packages/components/src/Workflow/_docs/ProgressStepper.stickersheet.stories.tsx
@@ -4,7 +4,7 @@ import { StickerSheet, StickerSheetStory } from '~storybook/components/StickerSh
 import { ProgressStepper } from '../subcomponents'
 
 export default {
-  title: 'Components/Workflow/Components/Progress Stepper',
+  title: 'Components/Workflow/Components/ProgressStepper',
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
