+++ packages/components/src/Workflow/_docs/ProgressStepper.stories.tsx
@@ -3,7 +3,7 @@ import { Workflow } from '../'
 import { WorkflowControls } from './controls'
 
 const meta = {
-  title: 'Components/Workflow/Components/Progress Stepper',
+  title: 'Components/Workflow/Components/ProgressStepper',
   component: Workflow.ProgressStepper,
   argTypes: {
     currentStepId: WorkflowControls.currentStepId,
